Split Hash trait
This is Kix's work from #1555, picking it up to help out. I've just
rebased it.

The `Hash` trait requires for the hash type to support hashing arbitrary
data. This is not great, since hash newtypes usually hash specific data.
Hashing arbitrary data into newtypes is usually a bug or
misunderstanding of the API.

This change separates methods performing hashing of arbitrary data to a separate RawHash trait and provides a few helpers to make usage easier:

- Add a private `engine()` method to the hash newtype
- Add a `finalize()` method to `HashEngine`
- Add a `hashes::hash` function

Moves `TxMerkleNode` hash to the `merkle_tree` module because that is
where implement the custom hashing and need access to its private inner
hash.

Closes #1481

Co-developed-by: Martin Habovstiak <martin.habovstiak@gmail.com>