hashes: Move ops trait bounds to associated Bytes type
The indexing ops trait bounds on `Hash` are a bit odd because hash is an
abstraction which contains an underlying byte array of some type. The
trait bounds are better put on the associated `Bytes` type rather than
on the type implementing `Hash`.

We only using indexing in `Hmac`, and can do so still by just calling
`as_byte_array()` before indexing.