absolute: Use InputString
In preparation for moving the absolute lock time height/time types to
the new `units` crate replace the use of `String` with `InputString`.

Note that in `ParseError` we are doing reasonable complex custom
displaying and this does not fit in that well with how the `InputString`
is designed so just use `Debug` inside the calls to `write!`.