internals: Add functionality to display raw InputString
The `InputString` is designed to add context around an input string so
that displaying it makes sense when we don't have "alloc" available and
there is no input string to print. However, when parsing absolute height
and time types we add our own context however it would still be useful
to use the `InputString` type because we want to move the absolute types
to `units` where we need a non-alloc version, which the `InputString`
provides. Since the `InputString` is an internal thing we can modify it
to fit our purposes - this does risk making the type even more
complicated to understand though.