bitcoin: Use no_std attribute
As we do for other crates; default to `no_std`.

The major issues with this is that we have, up until now, assumed we
have `extern crate std` in tests. Most of this patch is fixing the
feature gating to use stuff from `extern crate alloc` and correctly
feature gating on "std" if required.

Note also use of `format!` in tests instead of `println!` because it is
available in `alloc`.