chore: Hard code genesis script bytes instead of hex
Currently we have a dependency on `hex_lit` and it is used in exactly
one place outside of test code, if we instead use a hardcoded array
instead we can move the `hex_lit` dependency to be a dev-dependency
only.

Hard code the genesis block script bytes as an array, add a unit test
that parses the original hex so we know there were no mistakes.

Move the `hex_lit` dependency to the `dev-dependencies` of the manifest.