fix(CompactTarget): Manually implement comparison traits
The `CompactTarget` is a lossy encoding of a `Target`, as such two
`Target` objects may compare differently to the associated
`CompactTarget`, this is surprising.

We can manually implement all the comparison traits (incl. `Hash`) by
first converting the `CompactTarget` to a `Target` - this way any
comparison or ordering will be unsurprising.

Fix: #2110