bitcoin: Depend on and export hex-conservative
When we did the `rust-bitcoin v0.30.0` release we had not yet released
the `hex-conservative` crate so we forced users to reach into
`bitcoin-internals` to get at hex stuff e.g., `DisplayHex`.

To ease the upgrade path for any remaining upgraders add a dependency
on the now released `hex-conservative` crate and re-export it from the
crate root as `hex` (bit late I know, sorry).