bitcoin: Bump version to 0.32.0-rc1
In preparation for dropping the first release candidate bump the version
and add a changelog.

Please not I went to much more effort that usual with the changelog,
open to review on the overall form - not promising I'll change it but
definitely would like to keep iterating and improving.

If this changelog is appreciated then FWIW I don't think we should
bother automating it, a machine does not have all the context required
to create it.