test: Use 'm/' prefix in derivation path string
In the integration test of the new PSBT signing API the derivation path
strings do not have a leading 'm/'.

The PR that introduced the new API must have overlapped with the one
that removed the prefix from the `bip32` modules `DerivationPath` string
parsing because when we reverted that change these paths were not
included.