Slightly improve the usage of OutOfRangeError
Holy f**k, another case of an overly complicated error type stealing
time and brain power from me.

The `OutOfRangeError` is re-used to mean multiple things, it is complex
enough that we wrote a kani test that used it incorrectly, and it took
me an hour or so to debug it - awesome.

- Document the retarded usage of the `is_signed` field.
- Fix the broken kani test.
- Add a unit test to check no-one patches `OutOfRangeError::negative`
  function when they notice it looks insane.

Patch does not make the error type less complex because the whole
`amount` module is overly complicated and a real PITA to work on.