Remove Hash impl for new hashtypes
New hash types created with `hash_newtype` are not meant to be general
purpose hashes but currently we implement `Hash` for the newtype, this
makes the API for the newtype misleading.

Instead of implementing `Hash` we can make the `Hash` trait methods
inherent on the new type, this allows the module that creates the new
type to hash stuff ergonomically but does not pollute the public API for
the type.