Simplify the sha256t_hash_newtype macro
The generics on `sha256t::Hash` make working with the `hashes` module
more complicated that it needs to be. The only reason for the tag is to
have a way to keep a compile time midstate constant for the sha256
engine used by the tagged hash. Note also that the `sha25t::Hash` type
is only ever used in test code, any user of the `hashes` lib (including
`bitcoin` creates a new type and uses that).

The same can be achieved by creating new engine type and putting the
const midstate in the `Default` implementation.