Replace hash_newtype usage
Currently we use `hashes::hash_newtype!` to create various types that
wrap hash types. While this works it makes the `hashes` crate less
maintainable and more prone to change because logic that we want to
control from `bitcoin` is in `hashes`. We want to stabalise the leaf
crates as quickly as possible, moving the `hash_newtype` logic to
`bitcoin` allows us to simplify `hashes` and put the code closer to
where it is used.

Note, this patch does not remove `hashes::hash_newtype`, it just adds a
bunch of additional macros to `bitcoin::internal_macros`.

Note on design:

Instead of a single enormous macro, we add multiple smaller macros. This
allows implementing just what is needed for a type and gives finer
control over what is implemented and what is in the public API. All at
the cost of being a bit more verbose.