Replace sha256t_hash_newtype macro
As we did for the `hashes::hash_newtype!` macro; replace the
`hashes::sha256t_hash_newtype!` macro with a collection of more simple
macros in the `bitcoin` crate.

This improves maintainability by giving putting control of the macro
logic in the crate that uses the macro. Also assists in stabalizing
`hashes` sooner because we can now remove the two macros from `hashes`.