hashes: Remove both hash newtype macros
The `hashes` crate is _hard_ to work on, every time one wants to change
something there are various macros that are far from trivial to
understand.

We would like to improve the `hashes` API so we can stabalize the crate.
In an effort to assist in doing so remove the two hash newtype macros.

This is obviously an API breaking change - is anyone outside of
`rust-bitcoin` using these macros, if so they are going to be annoyed
with us.

Remove the `hash_newtype!` and `sha256_hash_newtype` macros.