hashes: Remove the sha256t module
As can be seen from the macros in `bitcoin` we can fully support tagged
SHA256 hashes without needing a module in `hashes`.

In an effort to make the `hashes` crate easier to work on remove the
`sha256t` module all together.

This risks rugging and annoying users but either I need 10 more points
of IQ to work on the `hashes` crate or we have to do something to reduce
the complexity.