Remove the sha256t::Hash type
In `bitcoin` we never use the `sha256t::Hash` type directly but only as
an underlying generic type that holds the tagging logic. To get tagged
hash types we currently use the `sha256t_hash_newtype` macro.

Assuming other users implement tagged hashing as we do in `rust-bitcoin`
then the `sha256t::Hash` type currently serves only as a type to
implement `Tag::engine` on - but as long as there exists a way to get
the pre-tagged engine then we do not actually need the type.

If we assume all users will use a provided macro to define tagged hashes
then we have two options. This PR implements the first option.

1. In the `NewTypeHash::engine` function return a pre-tagged
  `sha256::HashEngine`.

2. Have callers of the macro define a new engine type as well as a new
hash type, then we can put the tagging logic in the `Default`
implementation of the new engine.

Remove the `sha256t::Hash` type, and the `Tag` trait. Modify the
`sha256_hash_newtype` macro to match and just implement the tagging
logic in the `engine` function.

Add unit tests to make sure we correctly handle the forward/backward
attribute. `rust-bitcoin` includes test already that verify this patch.