CI: Add docs and document start
Add a document start and comment to help try to stop the readme from
going stale.

This removes a `yamllint` warning.

While we are clearing lint warnings disable the one causing

  warning  truthy value should be one of [false, true]  (truthy)

This is a know issue with GitHub actions yaml because `on` is a yaml
keyword, or something like that.