add `#![allow(unexpected_cfgs]` to every non-example target
As of Rust 1.80/nightly 2024-05-05, using custom `cfg` attributes
requires that we add a build.rs to all our crates. Aside from slowing
builds, this is a red flag for people trying to audit our crates because
prior to this change, there was no reason for a project to have a
build.rs except to interact with some other language -- which itself is
something that requires greater scrutiny than a normal Rust crate.

Obviously we are not going to do that for any of our crates except for
rust-secp, which wraps a C library.

We use custom cfg attributes to allow disabling our cryptography for
certain kinds of fuzztests. The Rust blog suggests using Cargo features
for this. This is even more insane, because it would then mean that if
*any dependency* in a user's tree were to enable a "disable all the
crypto" feature in our projects, the user would be fatally and
thoroughly compromised.

No doubt there will be a ton of pushback over the coming weeks but
bluntly, the Rust community is a dumpster fire and I have zero interest
in participating in this shit. There is an easy way to just disable the
new behavior so do that.