Move "private" public macros and hide docs
We have three macros that are only pubic because they are called from
other public macros:

- `hex_fmt_impl`
- `borrow_slice_impl`
- `serde_impl`

Thes macros are not meant to be used by library consumers. We have a
bunch of these types of macros and for all the others we hide them from
docs so as not to clutter the docs.

While we are at it put the two in `macros` down the bottom of the file
with the other "private" ones.