Add whitespace
The formatter lines up comments if they are on consecutive lines even
if the second is supposed to be at the start of the collum and the
first is after code. Putting a line of whitespace between the two
lines stops this from happening.

Add whitespace to stop the formatter doing silly changes.

Whitespace only.