hashes: Add regression tests
We have regression tests spread out throughout the `hashes` module but
they are not labelled as such. To give us more confidence and help
debug when patching the `hashes` crate we can add a bunch of regression
tests in a single place.

Add a module that does a single regression test for each type, simply
hash some arbitrary data and check the hex display against a hard coded
hex string.