Make Address:p2sh_from_hash public
We previously made this function Private and added a comment that doing
so was somehow better to remove the footgun of hashing the wrong length
script. However in hindsight this was a bad idea and users want the
functionality.

Make the `Address:p2sh_from_hash` public and document it as we do for
`Address::p2sh`.