fuzz: switch from serde_cbor to cborium
We were using an outdated CBOR crate for MSRV reasons. But in our fuzz
tests we have already given up on MSRV compatibility, and the outdated
crate was causing us problems. So replace it.

In general, I am suspicious of this decode-then-reencode test. CBOR has
some ambiguity in integer encoding. But empirically it has seemed to
work for a long time, and updating the crate seems to tighten the
decoder enough to get it working again, so we'll keep it for now.

Fixes #2801