Remove repetition for hash_newtype
Macros are hard enough to work on and read without adding the additional
complexity from repetition. Also, since the macro call site is
[hopefully] write-once-read-often code making the call site as quick to
read as possible is beneficial.

The changes in this patch are subjective, I'd argue that the real
benefit is in the future when we work more on refactoring and improving
the macro.