Introduce impl_bytelike_traits macro
As a step towards cleaning up the `hashes` macros introduce a new macro
that adds all the traits we want for byte-like types. Use it in
the `hash_trait_impls` macro.

This patch is essentially a refactor plus addition of `AsRef` and
`Borrow` for byte array (we have slice already).

Includes explicit update of the `hex-conservative` patch version
because we need the new `impl_fmt_traits` macro.