Link to std::error::Error
In #2521 I removed the link from `std::error::Error` with the claim that
it broke no-std builds. However there are a ton of other places where we
link to `std::` types.

I have no idea where the breakage was, I assume it existed and I was
sane at the time, CI on this patch will tell us.

Close: #2571